assignment2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
openai==1.3.0
azure-identity==1.14.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2